pymodels/requirements.txt
grpcio-tools

grpcio

scikit-learn
git+https://github.com/online-ml/river
